:truck: Rename TravelDeskFlightRequestsManageCard to TravelDeskFlightRequestsEditCard.
Why? To make space for a travel desk user version of the card.
